+++ lib/src/entities/local_document_model_response.dart
@@ -21,16 +21,16 @@ class LocalDocumentModelResponse {
       this.revisions});
 
   /// Holds offset where the document list started
-  int offset;
+  final int offset;
 
   /// List array of view row objects
-  List<Map<String, Object>> rows;
+  final List<Map<String, Object>> rows;
 
   /// Holds number of documents in the database
-  int totalRows;
+  final int totalRows;
 
   /// Current update sequence for the database
-  String updateSeq;
+  final String updateSeq;
 
   /// Holds local document object
   ///
@@ -48,35 +48,35 @@ class LocalDocumentModelResponse {
   /// This properties are listed separately in [LocalDocumentModelResponse] and you can get their directly.
   ///
   /// Returns by [LocalDocumentModel.localDoc]
-  Map<String, Object> doc;
+  final Map<String, Object> doc;
 
   /// Holds operation status. Available in case of success
-  bool ok;
+  final bool ok;
 
   /// Holds document ID
-  String id;
+  final String id;
 
   /// Holds revision info of document
-  String rev;
+  final String rev;
 
   /// Attachment's raw data
-  Object attachment;
+  final Object attachment;
 
   /// List of conflicted revisions
-  List<String> conflicts;
+  final List<String> conflicts;
 
   /// Deletion flag. Available if document was removed
-  bool deleted;
+  final bool deleted;
 
   /// List of deleted conflicted revisions
-  List<String> deletedConflicts;
+  final List<String> deletedConflicts;
 
   /// Document’s update sequence in current database
-  String localSeq;
+  final String localSeq;
 
   /// List of objects with information about local revisions and their status
-  List<Map<String, Object>> revsInfo;
+  final List<Map<String, Object>> revsInfo;
 
   /// List of local revision tokens without
-  Map<String, Object> revisions;
+  final Map<String, Object> revisions;
 }